requirements.txt
Flask==2.0.3
PyMySQL==1.0.2
mysql-connector-python==9.2.0
Werkzeug==2.0.3
cryptography==36.0.0
python-dotenv==1.0.1